requirements.txt
# Idea Spark Backend Dependencies

# Core Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database (PostgreSQL + pgvector)
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
pgvector==0.2.4

# AI/ML & RAG
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
tiktoken==0.5.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
bcrypt==4.1.2
email-validator==2.1.0

# File Processing
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
Pillow==10.1.0
python-magic==0.4.27

# Azure Services
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.26.0
aiofiles==23.2.0

# Monitoring & Logging
structlog==23.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
isort==5.13.2